認証済みページと未認証ページのルートグループ分離とヘッダー実装
- Next.jsのルートグループ機能を使って (private) と (public) に分離
- (private) グループに共通ヘッダーレイアウトを実装
  - ヘッダーにメニューボタンとユーザーボタンを配置
  - スクロール可能なメインコンテンツエリア
- 顧客一覧ページのUIを改善
  - タイトルを独立表示
  - 検索フォームとテーブルを別々のカードに配置
- グローバルCSSからscrollbar-gutter設定を削除し、必要な箇所のみに適用
- package.jsonにdevスクリプトを追加（Supabase起動 + Next.js起動）
