+++ apps/web/app/(public)/login/login-form.tsx
@@ -14,8 +14,8 @@ import { Input } from "@workspace/ui/components/input";
 import { Loader2 } from "lucide-react";
 import { useTransition } from "react";
 import { useForm } from "react-hook-form";
-import { type LoginSchema, loginSchema } from "../../features/auth/schema";
-import { useIsHydrated } from "../../libs/use-is-hydrated";
+import { type LoginSchema, loginSchema } from "../../../features/auth/schema";
+import { useIsHydrated } from "../../../libs/use-is-hydrated";
 import { loginAction } from "./login-action";
 
 export function LoginForm() {
