+++ package.json
@@ -6,6 +6,7 @@
 	},
 	"packageManager": "pnpm@10.12.4",
 	"scripts": {
+		"dev": "pnpm --filter @workspace/supabase supabase:start && pnpm --filter web dev",
 		"test:browser": "pnpm --filter web test:browser",
 		"test:e2e": "pnpm --filter @workspace/db db:reset && pnpm --filter web test:e2e",
 		"test:server": "pnpm --filter web test:server",
