+++ packages/ui/src/styles/globals.css
@@ -120,9 +120,6 @@
 	* {
 		@apply border-border outline-ring/50;
 	}
-	html {
-		scrollbar-gutter: stable;
-	}
 	body {
 		@apply bg-background text-foreground;
 	}
